backend/user-service/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
PyJWT==2.8.0
psycopg2-binary==2.9.9
redis==5.0.1
requests==2.31.0
